Move collector and processor to better dir
This makes the naming of the packages more consistent with the
diagrams in the design doc.
